requirements.txt
google-cloud-compute==0.5.0
gcloud-config-helper==0.3.0
boto3==1.18.36
pyyaml=5.4.1
